Reduce default buffer size
The OS is already doing file page buffering and bigger values provide
only marginal benefits. This is supported by the fact the tests show no
noticeable difference in execution time.

As noted in the readme, small numbers for the buffer size are good for
utilizing the available file size limit. Especially if there are a lot
of language variants. E.g. for 10 languages each buffer entry consists
of 130 = 10⋅(10+3) lines. So the previous buffer size of 1000 entries
might lead to a buffer which is already over 10 MiB big. In that case
the size limit is broken.